Implemented auto-close all windows on exit.
Implemented the ability to automatically close all the windows on
termination of AirWM. A method was added for the screen object to close
all the windows on a certain screen. In the main file it the
functionality was added by going through all the screens and calling the
new method.

Signed-off-by: Martin Jorn Rogalla <martin@martinrogalla.com>