Check redirect-override flag before adding window.
Before managing a window, check if the redirect-override falg is set. If
this is the case, map the window directly and do not manage the window.
This is needed for small popups such as the firefox search history
popup. Otherwise for each of these popups a full window would be
created.

Signed-off-by: Martin Jorn Rogalla <martin@martinrogalla.com>